refactor(common): build to dist & remove color.js
