+++ components/common/.gitignore
@@ -1,0 +1,9 @@
+dist/
+node_modules/
+
+.DS_Store
+
+# lock
+yarn.lock
+package-lock.json
+pnpm-lock.yaml
