+++ components/common/package.json
@@ -1,26 +1,33 @@
 {
   "name": "@omiu/common",
-  "version": "0.0.11",
+  "version": "0.0.12",
   "description": "Omi UI Components.",
-  "main": "src/index.js",
-  "module": "src/index.esm.js",
-  "scripts": {
-    "build": "rollup -c scripts/rollup.config.js && rollup -c scripts/rollup.config.umd.js"
+  "main": "dist/index.js",
+  "module": "dist/index.esm.js",
+  "types": "dist/index.d.ts",
+  "exports": {
+    ".": "./dist/index.esm.js",
+    "theme": "./scss/theme.scss",
+    "var": "./scss/var.scss"
   },
-  "keywords": [],
+  "scripts": {
+    "build": "rollup -c rollup.config.js",
+    "clean": "rm -rf dist",
+    "prepublishOnly": "npm run build"
+  },
+  "keywords": [
+    "omi"
+  ],
   "author": "dntzhang",
   "license": "MIT",
   "devDependencies": {
-    "@rollup/plugin-commonjs": "^11.1.0",
-    "@rollup/plugin-node-resolve": "^13.0.0",
-    "@rollup/plugin-typescript": "^8.2.1",
+    "@rollup/plugin-typescript": "8.2.5",
     "rollup": "^2.7.1",
-    "rollup-plugin-license": "^2.0.0"
+    "rollup-plugin-license": "^2.0.0",
+    "tslib": "^2.3.1",
+    "typescript": "4.3.5"
   },
   "dependencies": {
-    "@ctrl/tinycolor": "^3.4.0",
-    "color-convert": "^2.0.1",
-    "color-name": "^1.1.4",
-    "color-string": "^1.5.3"
+    "@ctrl/tinycolor": "^3.4.0"
   }
 }
