+++ components/common/scss/theme.scss
